Add comment form and Post it to the server
Remove hard-coded comments from the .java file. Add a form for the user to add comments. Add buttons for reset and submit. Add styles to it so that it blends with the home page background. Add Comments link to the navbar for both index and gallery.html. After adding the comment make the server to redirect to the point where the user was, showing the updated comment.