Finishing touches for week-3's work
Add option for user to select the number of comments they want to see. Server working properly with Datastore. Comments shown in the order of newest first. If there are no comments then don't display the 'delete-all-button' and the 'settings-button' and display a 'no comments' message. And fixes from the previous code reviews.