Consider optional attendees.
Now the code loops through all the consecutive free slots when checking for the overlapping event. Removed few methods as suggested in the code review. Changed variable names that contained '_' to according to the style guide. Add new methods that calculate free time slots for mandatory attendees and optional attendees. Using LinkedList instead of ArrayList. Use iterator to add or remove or set elements in the list while looping throught it.